# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MDM Platform API
  description: REST API for the MDM admin console. Manages devices, policies, apps, compliance, and more.
  version: 1.0.0
  contact:
    name: MDM Platform Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.staging.mdm.example.com/api/v1
    description: Staging
  - url: https://api.mdm.example.com/api/v1
    description: Production
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    ApiError:
      type: object
      required: [statusCode, error, message]
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
        details:
          type: object
        requestId:
          type: string
    Platform:
      type: string
      enum: [ios, android, windows, macos, linux]
    DeviceStatus:
      type: string
      enum: [enrolled, pending, blocked, wiped, retired]
    ComplianceStatus:
      type: string
      enum: [compliant, non_compliant, unknown]
...
    AppType:
      type: string
      enum: [public, enterprise, web_clip]

    # --- Auth ---
...
    Device:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        udid:
          type: string
        serialNumber:
          type: string
        name:
          type: string
        platform:
          $ref: '#/components/schemas/Platform'
        osVersion:
          type: string
        model:
          type: string
        manufacturer:
          type: string
        status:
          $ref: '#/components/schemas/DeviceStatus'
        complianceStatus:
          $ref: '#/components/schemas/ComplianceStatus'
        lastSeenAt:
          type: string
          format: date-time
        enrolledAt:
          type: string
          format: date-time
        deviceInfo:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    App:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        name:
          type: string
        bundleId:
          type: string
        platform:
          $ref: '#/components/schemas/Platform'
        type:
          $ref: '#/components/schemas/AppType'
        iconUrl:
          type: string
          format: uri
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAppRequest:
      type: object
      required: [name, bundleId, platform]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        bundleId:
          type: string
          minLength: 1
          maxLength: 255
        platform:
          $ref: '#/components/schemas/Platform'
        type:
          $ref: '#/components/schemas/AppType'
        iconUrl:
          type: string
          format: uri
        description:
          type: string
    UpdateAppRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        iconUrl:
          type: string
          format: uri
        description:
          type: string
...
    AppListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/App'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # --- Kiosk Profile ---
...
    Webhook:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateWebhookRequest:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
          minItems: 1
        status:
          type: string
          enum: [active, inactive]
          default: active
    UpdateWebhookRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive]
    WebhookWithSecret:
      allOf:
        - $ref: '#/components/schemas/Webhook'
        - type: object
          properties:
            secret:
              type: string
...
    WebhookListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Webhook'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NoContent:
      description: Successful deletion, no content returned
paths:
...
  /devices/{id}:
    get:
      operationId: getDevice
      summary: Get device by ID
      tags: [devices]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Device details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /devices/{id}/apps:
    get:
      operationId: getDeviceApps
      summary: Get apps assigned to a device
      tags: [devices]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Assigned apps
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        assignmentId:
                          type: string
                          format: uuid
                        appId:
                          type: string
                          format: uuid
                        appName:
                          type: string
                        bundleId:
                          type: string
                        isRequired:
                          type: boolean
                        assignedAt:
                          type: string
                          format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
...
  /apps:
    get:
      operationId: listApps
      summary: List apps
      tags: [apps]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of apps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppListResponse'
    post:
      operationId: createApp
      summary: Create an app
      tags: [apps]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppRequest'
      responses:
        '201':
          description: App created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App'
  /apps/{id}:
    get:
      operationId: getApp
      summary: Get app by ID
      tags: [apps]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: App details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateApp
      summary: Update an app
      tags: [apps]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppRequest'
      responses:
        '200':
          description: App updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteApp
      summary: Delete an app
      tags: [apps]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
...
  /webhooks:
    get:
      operationId: listWebhooks
      summary: List webhooks
      tags: [webhooks]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of webhooks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookListResponse'
    post:
      operationId: createWebhook
      summary: Create a webhook
      tags: [webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Webhook created (includes secret, shown only once)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookWithSecret'
  /webhooks/{id}:
    get:
      operationId: getWebhook
      summary: Get webhook by ID
      tags: [webhooks]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Webhook details (secret not included)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateWebhook
      summary: Update a webhook
      tags: [webhooks]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookRequest'
      responses:
        '200':
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteWebhook
      summary: Delete a webhook
      tags: [webhooks]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
...